requirements.txt
torch~=1.13.0
torchvision~=0.14.0
pillow>=9.2.0
requests~=2.28.1